Add CLI area progress bar helper (#25208)
